Fix question is not displayed
On the branch feature/ilub for ilias6, the questions are not displayed
in the presenters view while a live voting is running. The reason for
this is how the html-container for the question is set in the
main-template while creating the page. The template variable
'PLAYER_CONTENT' does not exist in the main-template (see code before my
commit). Therefore, the html-element, in which the question would be
loaded over ajax, does not exist. If the page content is set like in
older ilias versions (see the change of my commit), it works again. The
only the which looks a little bit odd after this change is the textblock
with how many people have voted.